agentcore/pyproject.toml
[project]
name = "agentcore"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bedrock-agentcore>=1.2.0",
    "strands-agents>=1.24.0",
    "python-dotenv>=1.0.0",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.0",
    "boto3>=1.35.0",
]
